+++ terraform/modules/gateway-api-integrator/variables.tf
@@ -5,16 +5,24 @@ variable "app_name" {
   description = "Name of the application in the Juju model."
   type        = string
   default     = "gateway-api-integrator"
+  nullable    = false
+}
+
+variable "base" {
+  description = "The operating system on which to deploy. null lets the provider use the charm's default."
+  type        = string
+  default     = null
 }
 
 variable "channel" {
   description = "The channel to use when deploying a charm."
   type        = string
   default     = "1/stable"
+  nullable    = false
 }
 
 variable "config" {
-  description = "Application config."
+  description = "Application config. Details about available options can be found at https://charmhub.io/gateway-api-integrator/configurations."
   type        = map(string)
   default     = {}
 }
@@ -22,28 +30,23 @@ variable "config" {
 variable "constraints" {
   description = "Juju constraints to apply for this application."
   type        = string
-  default     = "arch=amd64"
+  default     = null
 }
 
 variable "model_uuid" {
-  description = "Reference to a juju model's uuid."
+  description = "UUID of the Juju model where the application will be deployed."
   type        = string
+  nullable    = false
 }
 
 variable "revision" {
-  description = "Revision number of the charm"
+  description = "Revision number of the charm. null deploys the latest revision on the channel."
   type        = number
   default     = null
 }
 
-variable "base" {
-  description = "The operating system on which to deploy"
-  type        = string
-  default     = "ubuntu@24.04"
-}
-
 variable "units" {
-  description = "Number of units to deploy"
+  description = "Number of units to deploy."
   type        = number
   default     = 1
 }